Split monolithic provider into clean services architecture with dependency injection
